+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - Added commune ai-related  Validator API endpoints.
+- Add endpoint for returning available miners per network.
 
 
 ## [1.0.0] - 2024-10-23